Create Readme.md
In this repo are demonstrations of interesting coding examples using C#, VB, JS, .NET and SQL.
I made up the example to show the code. Please take the underlying sample concept with a grain of salt.
You might also say there's a shorter way to do it. There may well be. I'm simply experimenting with the constructs.